app-portage/eclass-manpages: Add @SUPPORTED_EAPIS tag for eclass
Add a @SUPPORTED_EAPIS tag that can be used to explicitly provide a list
of EAPIs that are supported by the eclass. The main goal is to make it
possible to extract this list with relative ease, for scripting
purposes. It is not included explicitly in the manpages at the moment.

The first use case is to make it possible to explicitly distinguish
eclasses that do not support a specific EAPI from eclasses that are not
used by any ebuilds using a specific EAPI. Therefore, it will make it
possible to easily detect when we can deprecate old EAPIs from eclasses.